Display DWM Window Mica Backdrop in Windows platform
