+++ example/android/src/main/kotlin/com/konyaco/fluent/example/MainActivity.kt
@@ -8,7 +8,9 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            App()
+            ExampleTheme {
+                App()
+            }
         }
     }
 }
