+++ example/desktop/src/jvmMain/kotlin/com/konyaco/fluent/example/Main.kt
@@ -5,6 +5,9 @@ import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.WindowPosition
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
+import com.mayakapps.compose.windowstyler.WindowBackdrop
+import com.mayakapps.compose.windowstyler.WindowStyle
+import org.jetbrains.skiko.hostOs
 
 fun main() = application {
     Window(
@@ -12,6 +15,12 @@ fun main() = application {
         state = rememberWindowState(position = WindowPosition(Alignment.Center)),
         title = "Compose Fluent"
     ) {
-        App()
+        ExampleTheme(displayMicaLayer = !hostOs.isWindows) {
+            WindowStyle(
+                isDarkTheme = LocalStore.current.darkMode,
+                backdropType = WindowBackdrop.Mica
+            )
+            App()
+        }
     }
 }
